[FIX] Fixed crash when image succeeds specific size (in pixels)
The bug was caused due to the limitation of a maximum of 4094 different XFs
(styles) in the xlwt-module. The feature which introduced the possibility to
add grids to the Excel sheet created a new style for every pixel.

This behaviour is now changed: The different style-variations (no border, top
border, left border, and top & left border) are now generated in advance and
re-used whenever they are needed.

Tested with images of various sizes (up to 1600x900) and different grid
settings.
Tested on Linux Mint 17.1 (64-bit) and LibreOffice 4.2.8.2