[TASK] Parse command line arguments with argparse module
In order to 1) clean up the codebase a bit and 2) make it easier to add new
features in the future, parsing command line arguments is now done with the
argparse module.

Unfortunately it is not possible to add new lines to the help-messages,
therefore '|' is used to separate explanations of the different format-choices.
Also, apparently it is not possible to rename the additional parameters for the
"--grid" option.

This commit also removes all functions which are not used anymore because of
this addition (print_usage, abort_with_usage, etc.). Hopefully I didn't forget
any.

Tested on Linux Mint 17.1 Cinnamon 64-bit and LibreOffice 4.2.8.2