[FIX] Fixed crash when running without --grid option
Cause of the bug was that in case the --grid option was not specified, the
variable args.grid was assigned the integer value 0. Later args.grid was
accessed as if it was an array.

The fix: In case --grid is not specified, [0, 0] (an array consisting of two
zeros) is now assigned to args.grid.

Tested on Linux Mint 17.1 Cinnamon 64-bit and LibreOffice 4.2.8.2